fix: move configs to config folder to avoid circular imports
