+++ TTS/tts/models/delightful_tts.py
@@ -1,7 +1,6 @@
 import logging
 import os
 from collections.abc import Callable
-from dataclasses import dataclass, field
 from itertools import chain
 from pathlib import Path
 from typing import Any
@@ -16,7 +15,7 @@ from torch.utils.data.sampler import WeightedRandomSampler
 from trainer.torch import DistributedSampler, DistributedSamplerWrapper
 from trainer.trainer_utils import get_optimizer, get_scheduler
 
-from TTS.config.shared_configs import ModelArgs
+from TTS.tts.configs.delightful_tts_config import DelightfulTTSConfig
 from TTS.tts.configs.shared_configs import BaseTTSConfig
 from TTS.tts.datasets.dataset import F0Dataset, TTSDataset, _parse_sample, get_attribute_balancer_weights
 from TTS.tts.layers.delightful_tts.acoustic_model import AcousticModel
@@ -270,103 +269,6 @@ class ForwardTTSE2eDataset(TTSDataset):
 
 
 ##############################
-# CONFIG DEFINITIONS
-##############################
-
-
-@dataclass
-class VocoderConfig(Coqpit):
-    resblock_type_decoder: str = "1"
-    resblock_kernel_sizes_decoder: list[int] = field(default_factory=lambda: [3, 7, 11])
-    resblock_dilation_sizes_decoder: list[list[int]] = field(default_factory=lambda: [[1, 3, 5], [1, 3, 5], [1, 3, 5]])
-    upsample_rates_decoder: list[int] = field(default_factory=lambda: [8, 8, 2, 2])
-    upsample_initial_channel_decoder: int = 512
-    upsample_kernel_sizes_decoder: list[int] = field(default_factory=lambda: [16, 16, 4, 4])
-    use_spectral_norm_discriminator: bool = False
-    upsampling_rates_discriminator: list[int] = field(default_factory=lambda: [4, 4, 4, 4])
-    periods_discriminator: list[int] = field(default_factory=lambda: [2, 3, 5, 7, 11])
-    pretrained_model_path: str | None = None
-
-
-@dataclass
-class DelightfulTtsAudioConfig(Coqpit):
-    sample_rate: int = 22050
-    hop_length: int = 256
-    win_length: int = 1024
-    fft_size: int = 1024
-    mel_fmin: float = 0.0
-    mel_fmax: float = 8000
-    num_mels: int = 100
-    pitch_fmax: float = 640.0
-    pitch_fmin: float = 1.0
-    resample: bool = False
-    preemphasis: float = 0.0
-    ref_level_db: int = 20
-    do_sound_norm: bool = False
-    log_func: str = "np.log10"
-    do_trim_silence: bool = True
-    trim_db: int = 45
-    do_rms_norm: bool = False
-    db_level: float = None
-    power: float = 1.5
-    griffin_lim_iters: int = 60
-    spec_gain: int = 20
-    do_amp_to_db_linear: bool = True
-    do_amp_to_db_mel: bool = True
-    min_level_db: int = -100
-    max_norm: float = 4.0
-
-
-@dataclass
-class DelightfulTtsArgs(ModelArgs):
-    num_chars: int = 100
-    spec_segment_size: int = 32
-    n_hidden_conformer_encoder: int = 512
-    n_layers_conformer_encoder: int = 6
-    n_heads_conformer_encoder: int = 8
-    dropout_conformer_encoder: float = 0.1
-    kernel_size_conv_mod_conformer_encoder: int = 7
-    kernel_size_depthwise_conformer_encoder: int = 7
-    lrelu_slope: float = 0.3
-    n_hidden_conformer_decoder: int = 512
-    n_layers_conformer_decoder: int = 6
-    n_heads_conformer_decoder: int = 8
-    dropout_conformer_decoder: float = 0.1
-    kernel_size_conv_mod_conformer_decoder: int = 11
-    kernel_size_depthwise_conformer_decoder: int = 11
-    bottleneck_size_p_reference_encoder: int = 4
-    bottleneck_size_u_reference_encoder: int = 512
-    ref_enc_filters_reference_encoder = [32, 32, 64, 64, 128, 128]
-    ref_enc_size_reference_encoder: int = 3
-    ref_enc_strides_reference_encoder = [1, 2, 1, 2, 1]
-    ref_enc_pad_reference_encoder = [1, 1]
-    ref_enc_gru_size_reference_encoder: int = 32
-    ref_attention_dropout_reference_encoder: float = 0.2
-    token_num_reference_encoder: int = 32
-    predictor_kernel_size_reference_encoder: int = 5
-    n_hidden_variance_adaptor: int = 512
-    kernel_size_variance_adaptor: int = 5
-    dropout_variance_adaptor: float = 0.5
-    n_bins_variance_adaptor: int = 256
-    emb_kernel_size_variance_adaptor: int = 3
-    use_speaker_embedding: bool = False
-    num_speakers: int = 0
-    speakers_file: str = None
-    d_vector_file: str = None
-    speaker_embedding_channels: int = 384
-    use_d_vector_file: bool = False
-    d_vector_dim: int = 0
-    freeze_vocoder: bool = False
-    freeze_text_encoder: bool = False
-    freeze_duration_predictor: bool = False
-    freeze_pitch_predictor: bool = False
-    freeze_energy_predictor: bool = False
-    freeze_basis_vectors_predictor: bool = False
-    freeze_decoder: bool = False
-    length_scale: float = 1.0
-
-
-##############################
 # MODEL DEFINITION
 ##############################
 class DelightfulTTS(BaseTTSE2E):
@@ -402,7 +304,7 @@ class DelightfulTTS(BaseTTSE2E):
     # pylint: disable=dangerous-default-value
     def __init__(
         self,
-        config: Coqpit,
+        config: DelightfulTTSConfig,
         ap,
         tokenizer: "TTSTokenizer" = None,
         speaker_manager: SpeakerManager = None,
