+++ TTS/tts/models/forward_tts.py
@@ -1,12 +1,10 @@
 import logging
-from dataclasses import dataclass, field
 
 import torch
 from coqpit import Coqpit
 from monotonic_alignment_search import maximum_path
 from torch import nn
 
-from TTS.config.shared_configs import ModelArgs
 from TTS.tts.layers.feed_forward.decoder import Decoder
 from TTS.tts.layers.feed_forward.encoder import Encoder
 from TTS.tts.layers.generic.aligner import AlignmentNetwork
@@ -19,156 +17,6 @@ from TTS.tts.utils.text.tokenizer import TTSTokenizer
 from TTS.tts.utils.visual import plot_alignment, plot_avg_energy, plot_avg_pitch, plot_spectrogram
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class ForwardTTSArgs(ModelArgs):
-    """ForwardTTS Model arguments.
-
-    Args:
-
-        num_chars (int):
-            Number of characters in the vocabulary. Defaults to 100.
-
-        out_channels (int):
-            Number of output channels. Defaults to 80.
-
-        hidden_channels (int):
-            Number of base hidden channels of the model. Defaults to 512.
-
-        use_aligner (bool):
-            Whether to use aligner network to learn the text to speech alignment or use pre-computed durations.
-            If set False, durations should be computed by `TTS/bin/compute_attention_masks.py` and path to the
-            pre-computed durations must be provided to `config.datasets[0].meta_file_attn_mask`. Defaults to True.
-
-        use_pitch (bool):
-            Use pitch predictor to learn the pitch. Defaults to True.
-
-        use_energy (bool):
-            Use energy predictor to learn the energy. Defaults to True.
-
-        duration_predictor_hidden_channels (int):
-            Number of hidden channels in the duration predictor. Defaults to 256.
-
-        duration_predictor_dropout_p (float):
-            Dropout rate for the duration predictor. Defaults to 0.1.
-
-        duration_predictor_kernel_size (int):
-            Kernel size of conv layers in the duration predictor. Defaults to 3.
-
-        pitch_predictor_hidden_channels (int):
-            Number of hidden channels in the pitch predictor. Defaults to 256.
-
-        pitch_predictor_dropout_p (float):
-            Dropout rate for the pitch predictor. Defaults to 0.1.
-
-        pitch_predictor_kernel_size (int):
-            Kernel size of conv layers in the pitch predictor. Defaults to 3.
-
-        pitch_embedding_kernel_size (int):
-            Kernel size of the projection layer in the pitch predictor. Defaults to 3.
-
-        energy_predictor_hidden_channels (int):
-            Number of hidden channels in the energy predictor. Defaults to 256.
-
-        energy_predictor_dropout_p (float):
-            Dropout rate for the energy predictor. Defaults to 0.1.
-
-        energy_predictor_kernel_size (int):
-            Kernel size of conv layers in the energy predictor. Defaults to 3.
-
-        energy_embedding_kernel_size (int):
-            Kernel size of the projection layer in the energy predictor. Defaults to 3.
-
-        positional_encoding (bool):
-            Whether to use positional encoding. Defaults to True.
-
-        positional_encoding_use_scale (bool):
-            Whether to use a learnable scale coeff in the positional encoding. Defaults to True.
-
-        length_scale (int):
-            Length scale that multiplies the predicted durations. Larger values result slower speech. Defaults to 1.0.
-
-        encoder_type (str):
-            Type of the encoder module. One of the encoders available in :class:`TTS.tts.layers.feed_forward.encoder`.
-            Defaults to `fftransformer` as in the paper.
-
-        encoder_params (dict):
-            Parameters of the encoder module. Defaults to ```{"hidden_channels_ffn": 1024, "num_heads": 1, "num_layers": 6, "dropout_p": 0.1}```
-
-        decoder_type (str):
-            Type of the decoder module. One of the decoders available in :class:`TTS.tts.layers.feed_forward.decoder`.
-            Defaults to `fftransformer` as in the paper.
-
-        decoder_params (str):
-            Parameters of the decoder module. Defaults to ```{"hidden_channels_ffn": 1024, "num_heads": 1, "num_layers": 6, "dropout_p": 0.1}```
-
-        detach_duration_predictor (bool):
-            Detach the input to the duration predictor from the earlier computation graph so that the duraiton loss
-            does not pass to the earlier layers. Defaults to True.
-
-        max_duration (int):
-            Maximum duration accepted by the model. Defaults to 75.
-
-        num_speakers (int):
-            Number of speakers for the speaker embedding layer. Defaults to 0.
-
-        speakers_file (str):
-            Path to the speaker mapping file for the Speaker Manager. Defaults to None.
-
-        speaker_embedding_channels (int):
-            Number of speaker embedding channels. Defaults to 256.
-
-        use_d_vector_file (bool):
-            Enable/Disable the use of d-vectors for multi-speaker training. Defaults to False.
-
-        d_vector_dim (int):
-            Number of d-vector channels. Defaults to 0.
-
-    """
-
-    num_chars: int = None
-    out_channels: int = 80
-    hidden_channels: int = 384
-    use_aligner: bool = True
-    # pitch params
-    use_pitch: bool = True
-    pitch_predictor_hidden_channels: int = 256
-    pitch_predictor_kernel_size: int = 3
-    pitch_predictor_dropout_p: float = 0.1
-    pitch_embedding_kernel_size: int = 3
-
-    # energy params
-    use_energy: bool = False
-    energy_predictor_hidden_channels: int = 256
-    energy_predictor_kernel_size: int = 3
-    energy_predictor_dropout_p: float = 0.1
-    energy_embedding_kernel_size: int = 3
-
-    # duration params
-    duration_predictor_hidden_channels: int = 256
-    duration_predictor_kernel_size: int = 3
-    duration_predictor_dropout_p: float = 0.1
-
-    positional_encoding: bool = True
-    poisitonal_encoding_use_scale: bool = True
-    length_scale: int = 1
-    encoder_type: str = "fftransformer"
-    encoder_params: dict = field(
-        default_factory=lambda: {"hidden_channels_ffn": 1024, "num_heads": 1, "num_layers": 6, "dropout_p": 0.1}
-    )
-    decoder_type: str = "fftransformer"
-    decoder_params: dict = field(
-        default_factory=lambda: {"hidden_channels_ffn": 1024, "num_heads": 1, "num_layers": 6, "dropout_p": 0.1}
-    )
-    detach_duration_predictor: bool = False
-    max_duration: int = 75
-    num_speakers: int = 1
-    use_speaker_embedding: bool = False
-    speakers_file: str = None
-    use_d_vector_file: bool = False
-    d_vector_dim: int = None
-    d_vector_file: str = None
 
 
 class ForwardTTS(BaseTTS):
@@ -194,8 +42,9 @@ class ForwardTTS(BaseTTS):
             Defaults to None.
 
     Examples:
-        >>> from TTS.tts.models.fast_pitch import ForwardTTS, ForwardTTSArgs
-        >>> config = ForwardTTSArgs()
+        >>> from TTS.tts.configs.fast_pitch_config import FastPitchConfig
+        >>> from TTS.tts.models.fast_pitch import ForwardTTS
+        >>> config = FastPitchConfig()
         >>> model = ForwardTTS(config)
     """
