+++ TTS/tts/models/tortoise.py
@@ -1,18 +1,16 @@
 import logging
 import os
 import random
-from dataclasses import dataclass
 from time import time
 from typing import Any
 
 import torch
 import torch.nn.functional as F
 import torchaudio
-from coqpit import Coqpit
 from tqdm import tqdm
 
-from TTS.config.shared_configs import ModelArgs
 from TTS.tts.configs.shared_configs import BaseTTSConfig
+from TTS.tts.configs.tortoise_config import TortoiseConfig
 from TTS.tts.layers.tortoise.arch_utils import TorchMelSpectrogram
 from TTS.tts.layers.tortoise.audio_utils import (
     denormalize_tacotron_mel,
@@ -211,119 +209,6 @@ def pick_best_batch_size_for_gpu():
     return batch_size
 
 
-@dataclass
-class TortoiseAudioConfig(Coqpit):
-    sample_rate: int = 22050
-    diffusion_sample_rate: int = 24000
-    output_sample_rate: int = 24000
-
-
-@dataclass
-class TortoiseArgs(ModelArgs):
-    """A dataclass to represent Tortoise model arguments that define the model structure.
-
-    Args:
-        autoregressive_batch_size (int): The size of the auto-regressive batch.
-        enable_redaction (bool, optional): Whether to enable redaction. Defaults to True.
-        high_vram (bool, optional): Deprecated, has no effect.
-        kv_cache (bool, optional): Whether to use the kv_cache. Defaults to True.
-        ar_checkpoint (str, optional): The checkpoint for the autoregressive model. Defaults to None.
-        clvp_checkpoint (str, optional): The checkpoint for the ConditionalLatentVariablePerseq model. Defaults to None.
-        diff_checkpoint (str, optional): The checkpoint for the DiffTTS model. Defaults to None.
-        num_chars (int, optional): The maximum number of characters to generate. Defaults to 255.
-
-        For UnifiedVoice model:
-        ar_max_mel_tokens (int, optional): The maximum mel tokens for the autoregressive model. Defaults to 604.
-        ar_max_text_tokens (int, optional): The maximum text tokens for the autoregressive model. Defaults to 402.
-        ar_max_conditioning_inputs (int, optional): The maximum conditioning inputs for the autoregressive model. Defaults to 2.
-        ar_layers (int, optional): The number of layers for the autoregressive model. Defaults to 30.
-        ar_model_dim (int, optional): The model dimension for the autoregressive model. Defaults to 1024.
-        ar_heads (int, optional): The number of heads for the autoregressive model. Defaults to 16.
-        ar_number_text_tokens (int, optional): The number of text tokens for the autoregressive model. Defaults to 255.
-        ar_start_text_token (int, optional): The start text token for the autoregressive model. Defaults to 255.
-        ar_checkpointing (bool, optional): Whether to use checkpointing for the autoregressive model. Defaults to False.
-        ar_train_solo_embeddings (bool, optional): Whether to train embeddings for the autoregressive model. Defaults to False.
-
-        For DiffTTS model:
-        diff_model_channels (int, optional): The number of channels for the DiffTTS model. Defaults to 1024.
-        diff_num_layers (int, optional): The number of layers for the DiffTTS model. Defaults to 10.
-        diff_in_channels (int, optional): The input channels for the DiffTTS model. Defaults to 100.
-        diff_out_channels (int, optional): The output channels for the DiffTTS model. Defaults to 200.
-        diff_in_latent_channels (int, optional): The input latent channels for the DiffTTS model. Defaults to 1024.
-        diff_in_tokens (int, optional): The input tokens for the DiffTTS model. Defaults to 8193.
-        diff_dropout (int, optional): The dropout percentage for the DiffTTS model. Defaults to 0.
-        diff_use_fp16 (bool, optional): Whether to use fp16 for the DiffTTS model. Defaults to False.
-        diff_num_heads (int, optional): The number of heads for the DiffTTS model. Defaults to 16.
-        diff_layer_drop (int, optional): The layer dropout percentage for the DiffTTS model. Defaults to 0.
-        diff_unconditioned_percentage (int, optional): The percentage of unconditioned inputs for the DiffTTS model. Defaults to 0.
-
-        For ConditionalLatentVariablePerseq model:
-        clvp_dim_text (int): The dimension of the text input for the CLVP module. Defaults to 768.
-        clvp_dim_speech (int): The dimension of the speech input for the CLVP module. Defaults to 768.
-        clvp_dim_latent (int): The dimension of the latent representation for the CLVP module. Defaults to 768.
-        clvp_num_text_tokens (int): The number of text tokens used by the CLVP module. Defaults to 256.
-        clvp_text_enc_depth (int): The depth of the text encoder in the CLVP module. Defaults to 20.
-        clvp_text_seq_len (int): The maximum sequence length of the text input for the CLVP module. Defaults to 350.
-        clvp_text_heads (int): The number of attention heads used by the text encoder in the CLVP module. Defaults to 12.
-        clvp_num_speech_tokens (int): The number of speech tokens used by the CLVP module. Defaults to 8192.
-        clvp_speech_enc_depth (int): The depth of the speech encoder in the CLVP module. Defaults to 20.
-        clvp_speech_heads (int): The number of attention heads used by the speech encoder in the CLVP module. Defaults to 12.
-        clvp_speech_seq_len (int): The maximum sequence length of the speech input for the CLVP module. Defaults to 430.
-        clvp_use_xformers (bool): A flag indicating whether the model uses transformers in the CLVP module. Defaults to True.
-        duration_const (int): A constant value used in the model. Defaults to 102400.
-    """
-
-    autoregressive_batch_size: int = 1
-    enable_redaction: bool = False
-    high_vram: bool = False
-    kv_cache: bool = True
-    ar_checkpoint: str = None
-    clvp_checkpoint: str = None
-    diff_checkpoint: str = None
-    num_chars: int = 255
-
-    # UnifiedVoice params
-    ar_max_mel_tokens: int = 604
-    ar_max_text_tokens: int = 402
-    ar_max_conditioning_inputs: int = 2
-    ar_layers: int = 30
-    ar_model_dim: int = 1024
-    ar_heads: int = 16
-    ar_number_text_tokens: int = 255
-    ar_start_text_token: int = 255
-    ar_checkpointing: bool = False
-    ar_train_solo_embeddings: bool = False
-
-    # DiffTTS params
-    diff_model_channels: int = 1024
-    diff_num_layers: int = 10
-    diff_in_channels: int = 100
-    diff_out_channels: int = 200
-    diff_in_latent_channels: int = 1024
-    diff_in_tokens: int = 8193
-    diff_dropout: int = 0
-    diff_use_fp16: bool = False
-    diff_num_heads: int = 16
-    diff_layer_drop: int = 0
-    diff_unconditioned_percentage: int = 0
-
-    # clvp params
-    clvp_dim_text: int = 768
-    clvp_dim_speech: int = 768
-    clvp_dim_latent: int = 768
-    clvp_num_text_tokens: int = 256
-    clvp_text_enc_depth: int = 20
-    clvp_text_seq_len: int = 350
-    clvp_text_heads: int = 12
-    clvp_num_speech_tokens: int = 8192
-    clvp_speech_enc_depth: int = 20
-    clvp_speech_heads: int = 12
-    clvp_speech_seq_len: int = 430
-    clvp_use_xformers: bool = True
-    # constants
-    duration_const: int = 102400
-
-
 class Tortoise(BaseTTS):
     """Tortoise model class.
 
@@ -337,7 +222,7 @@ class Tortoise(BaseTTS):
         >>> model.load_checkpoint(config, checkpoint_dir="paths/to/models_dir/", eval=True)
     """
 
-    def __init__(self, config: Coqpit):
+    def __init__(self, config: TortoiseConfig):
         super().__init__(config, ap=None, tokenizer=None)
         self.mel_norm_path = None
         self.config = config
