+++ TTS/tts/models/xtts.py
@@ -1,6 +1,5 @@
 import logging
 import os
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
@@ -11,7 +10,6 @@ import torchaudio
 from coqpit import Coqpit
 from trainer.io import load_fsspec
 
-from TTS.config.shared_configs import ModelArgs
 from TTS.tts.configs.shared_configs import BaseTTSConfig
 from TTS.tts.layers.xtts.gpt import GPT
 from TTS.tts.layers.xtts.hifigan_decoder import HifiDecoder
@@ -98,89 +96,6 @@ def load_audio(audiopath, sampling_rate):
     # clip audio invalid values
     audio.clip_(-1, 1)
     return audio
-
-
-@dataclass
-class XttsAudioConfig(Coqpit):
-    """
-    Configuration class for audio-related parameters in the XTTS model.
-
-    Args:
-        sample_rate (int): The sample rate in which the GPT operates.
-        output_sample_rate (int): The sample rate of the output audio waveform.
-        dvae_sample_rate (int): The sample rate of the DVAE
-    """
-
-    sample_rate: int = 22050
-    output_sample_rate: int = 24000
-    dvae_sample_rate: int = 22050
-
-
-@dataclass
-class XttsArgs(ModelArgs):
-    """A dataclass to represent XTTS model arguments that define the model structure.
-
-    Args:
-        gpt_batch_size (int): The size of the auto-regressive batch.
-        enable_redaction (bool, optional): Whether to enable redaction. Defaults to True.
-        kv_cache (bool, optional): Whether to use the kv_cache. Defaults to True.
-        gpt_checkpoint (str, optional): The checkpoint for the autoregressive model. Defaults to None.
-        clvp_checkpoint (str, optional): The checkpoint for the ConditionalLatentVariablePerseq model. Defaults to None.
-        decoder_checkpoint (str, optional): The checkpoint for the DiffTTS model. Defaults to None.
-        num_chars (int, optional): The maximum number of characters to generate. Defaults to 255.
-
-        For GPT model:
-        gpt_max_audio_tokens (int, optional): The maximum mel tokens for the autoregressive model. Defaults to 604.
-        gpt_max_text_tokens (int, optional): The maximum text tokens for the autoregressive model. Defaults to 402.
-        gpt_max_prompt_tokens (int, optional): The maximum prompt tokens or the autoregressive model. Defaults to 70.
-        gpt_layers (int, optional): The number of layers for the autoregressive model. Defaults to 30.
-        gpt_n_model_channels (int, optional): The model dimension for the autoregressive model. Defaults to 1024.
-        gpt_n_heads (int, optional): The number of heads for the autoregressive model. Defaults to 16.
-        gpt_number_text_tokens (int, optional): The number of text tokens for the autoregressive model. Defaults to 255.
-        gpt_start_text_token (int, optional): The start text token for the autoregressive model. Defaults to 255.
-        gpt_checkpointing (bool, optional): Whether to use checkpointing for the autoregressive model. Defaults to False.
-        gpt_train_solo_embeddings (bool, optional): Whether to train embeddings for the autoregressive model. Defaults to False.
-        gpt_code_stride_len (int, optional): The hop_size of dvae and consequently of the gpt output. Defaults to 1024.
-        gpt_use_masking_gt_prompt_approach (bool, optional):  If True, it will use ground truth as prompt and it will mask the loss to avoid repetition. Defaults to True.
-        gpt_use_perceiver_resampler (bool, optional):  If True, it will use perceiver resampler from flamingo paper - https://arxiv.org/abs/2204.14198. Defaults to False.
-    """
-
-    gpt_batch_size: int = 1
-    enable_redaction: bool = False
-    kv_cache: bool = True
-    gpt_checkpoint: str = None
-    clvp_checkpoint: str = None
-    decoder_checkpoint: str = None
-    num_chars: int = 255
-
-    # XTTS GPT Encoder params
-    tokenizer_file: str = ""
-    gpt_max_audio_tokens: int = 605
-    gpt_max_text_tokens: int = 402
-    gpt_max_prompt_tokens: int = 70
-    gpt_layers: int = 30
-    gpt_n_model_channels: int = 1024
-    gpt_n_heads: int = 16
-    gpt_number_text_tokens: int = None
-    gpt_start_text_token: int = None
-    gpt_stop_text_token: int = None
-    gpt_num_audio_tokens: int = 8194
-    gpt_start_audio_token: int = 8192
-    gpt_stop_audio_token: int = 8193
-    gpt_code_stride_len: int = 1024
-    gpt_use_masking_gt_prompt_approach: bool = True
-    gpt_use_perceiver_resampler: bool = False
-
-    # HifiGAN Decoder params
-    input_sample_rate: int = 22050
-    output_sample_rate: int = 24000
-    output_hop_length: int = 256
-    decoder_input_dim: int = 1024
-    d_vector_dim: int = 512
-    cond_d_vector_in_each_upsampling_layer: bool = True
-
-    # constants
-    duration_const: int = 102400
 
 
 class Xtts(BaseTTS):
