+++ recipes/ljspeech/vits_tts/train_vits.py
@@ -3,9 +3,9 @@ import os
 from trainer import Trainer, TrainerArgs
 
 from TTS.tts.configs.shared_configs import BaseDatasetConfig
-from TTS.tts.configs.vits_config import VitsConfig
+from TTS.tts.configs.vits_config import VitsAudioConfig, VitsConfig
 from TTS.tts.datasets import load_tts_samples
-from TTS.tts.models.vits import Vits, VitsAudioConfig
+from TTS.tts.models.vits import Vits
 from TTS.tts.utils.text.tokenizer import TTSTokenizer
 from TTS.utils.audio import AudioProcessor
